populate_base/requirements.txt
python-binance
elasticsearch
mysql-connector-python